requirements.txt

fastapi
uvicorn
passlib[bcrypt]
python-jose
yt-dlp
openai-whisper
